feat: add renet support to Unreal Engine
This commit adds a Renet plugin to provide easy integration with
the Rust Renet library in Unreal Engine. The plugin is written in C++ and Rust.
The Rust code generates a library that exports Renet's functions to C using FFI,
then the C module is imported and integrated with an Unreal Engine actor using the C++ code.

The plugin is designed to be used as a singleton, so the Rust part of the plugin
uses a global variable CLIENT. While this decision carries some technical debt,
it was made to simplify the library implementation, speed up development time,
and reduce potential errors with the incompatibility of the Rust FFI and Unreal Engine C++. T
he static const should be changed to a Box in the future.