refactor: create a new Unreal Engine project
We've created a new Unreal Engine project to provide better support for c++.
The previous template was designed to be used with Blueprints, but we have
decided to use c++ instead. This decision was made because we want to integrate
our project with our Renet UE5 plugin, which is intended to work with c++ projects.
This commit includes over 500 changes, all of which are related to the transition
from the Blueprint to the c++ template.