refactor: create basic Babylon.JS project
Create a basic Babylon.JS project (rotating cube + main menu UI), using SolidJS as the GUI backend.
The migration from Unreal Engine was necessary due to the lack of a good scripting
language and the challenges and limitations of development in C++.

Unity 3D wasn't an option due to issues with the UI Toolkit in its current state
and limited possibilities for extending the existing Unity app, requiring additional
work for creating DLC or mods.

Therefore, I decided to move the client-side of the project to TypeScript,
which will make the game more adaptable. Additionally, transitioning from C++/C# to JavaScript
shouldn't result in bad preformance since the server, written in Rust, will
handle all the heavy calculations.

JavaScript is also great from a UI perspective because it provides access to SolidJS and GSAP.