refactor: port server to WebSocket
Rewrite server implementation from RENet to WebSocket to simplify the client side of the project.
Previously, the client side required a complicated WebAssembly version of the RENet library.
Now, the client side can use the native-to-web and easy-to-use WebSocket protocol.

Most of the network code was moved into a separate crate called `bevy_websocket_server`
to simplify the project structure. The network plugin now only adds the `WebSocketServer`
plugin and handles Connect and Disconnect events.

The `channels` crate was removed as it was related to RENet, and the `message`
crate was merged with the `bevy_websocket_server` crate.