refactor: change Map's entity to a generic type
Change Bevy entity to a generic type to increase reusability and flexibility
of the resource.

While working on the Movement system, I realized that I needed to make multiple
changes to the Map's entity type in the resources.rs file. To simplify the process
and avoid making repetitive modifications, I made the entity type generic.