feat: add player movement system
Create resources, messages, and systems that handle the MoveTo request and
move the player to the requested position.

Create the `./map/components.rs` file that contains two components:
- `MapEntity` for storing the entity position and rotation.
- `MapEntityMoveTarget` for storing the target position.

Delete the `EntityMoved` event from the `./map/events.rs` file as it was never used.
Modify the `spawn_or_despawn_entity` system from the `./map/systems.rs` file to
delete the `MapEntity` and `MapEntityMoveTarget` components if the entity was removed
from the map, and add the `MapEntity` component when an entity was added to the map.

Create the `update_entity_position` system in the `./map/systems.rs` file
to update the position of all entities that contain the `MapEntity`
and `MapEntityMoveTarget` components.

Add the `MoveTo` message to the `./player/messages.rs` file and also create the
handler system for it in the `./player/systems.rs` (`request_move_player`) that
adds a new instance of the `MapEntityMoveTarget` component to the player entity.
This change will allow the client to request the player to move to a specific position.

Update the `register` system in the `./player/systems.rs` to add the new player
`EntitySpawned` component.

Add the `GetPlayerAdmiralId` message to the `./player/messages.rs` file and also
create the handler system for it in the `./player/systems.rs` (`request_admiral_id`) that
returns the ID of the client's admiral.